Ensure entire directory tree is created
In the event an intermediate directory does not exist, this command
fails. This ensures every intermediate directory is created
along the way, if it doesn't already exist.

This prevents errors if a user is following the instructions in
`documentation/Tophatting documentation.md`, but has never run `dev
up` in `Shopify/polaris-styleguide`.